# policy: check if rules already select endpoint in resolveL4{Ingress,Egress}Policy